[budget] Add real costs to the CSV export
When the real costs view is active, the exported CSV now mirrors the
on-screen columns: real cost per past month, real total, estimated so
far and gap, estimated future months, remaining, real + remaining,
estimated total and final gap.

The expense computations (converted expenses, done and remaining
previsional, extended departments) move from BudgetList to Budget so the
export and the table share the same data.
